microservices/alarm-processor/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
redis==5.0.1
pydantic==2.11.7
httpx==0.27.0
psycopg2-binary==2.9.10
pytz>=2023.3 